Adds e2e test.
